A Brief Survey on Anomaly Detection in Smart Grid
Smart grid is a powerful distributed energy network that performs a two-way communication between the consumers and the utilities. This can be considered as a complex system that interconnects various physical components like power plants and logical components like communication protocols and infrastructure. The main theme of smart grid is to distribute the electricity efficiently and acquire the data about electricity usage and transmit it back to the control center with the help of sensors attached to it. While the communication is in progress, some kind of irregularities like physical system faults or data hacking happens and results in delivering the wrong data. These kinds of anomalous behavior must be detected, else it deliver the wrong information and leads to make a wrong decision that can affect the efficiency of power distribution. In this paper, a survey on different types of attacks that can happen within the smart grid and proposed latest defense strategies that can protect the system from these attacks are explained in detail.